Fix closing and joining rooms
